// src/api/openapi/models/TransportRequirementStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TransportRequirementStatus = {
    DRAFT: 'Draft',
    PENDING_VERIFICATION: 'PendingVerification',
    VERIFIED: 'Verified',
    ARCHIVED: 'Archived',
    NOT_VERIFIED: 'NotVerified'
} as const;
export type TransportRequirementStatus = typeof TransportRequirementStatus[keyof typeof TransportRequirementStatus];


export function instanceOfTransportRequirementStatus(value: any): boolean {
    for (const key in TransportRequirementStatus) {
        if (Object.prototype.hasOwnProperty.call(TransportRequirementStatus, key)) {
            if (TransportRequirementStatus[key as keyof typeof TransportRequirementStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TransportRequirementStatusFromJSON(json: any): TransportRequirementStatus {
    return TransportRequirementStatusFromJSONTyped(json, false);
}

export function TransportRequirementStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransportRequirementStatus {
    return json as TransportRequirementStatus;
}

export function TransportRequirementStatusToJSON(value?: TransportRequirementStatus | null): any {
    return value as any;
}

export function TransportRequirementStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): TransportRequirementStatus {
    return value as TransportRequirementStatus;
}