// src/api/openapi/models/TransportRequirementsType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TransportRequirementsType = {
    CONCESSION: 'Concession',
    BUSINESS_RISK: 'Business_risk'
} as const;
export type TransportRequirementsType = typeof TransportRequirementsType[keyof typeof TransportRequirementsType];


export function instanceOfTransportRequirementsType(value: any): boolean {
    for (const key in TransportRequirementsType) {
        if (Object.prototype.hasOwnProperty.call(TransportRequirementsType, key)) {
            if (TransportRequirementsType[key as keyof typeof TransportRequirementsType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TransportRequirementsTypeFromJSON(json: any): TransportRequirementsType {
    return TransportRequirementsTypeFromJSONTyped(json, false);
}

export function TransportRequirementsTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransportRequirementsType {
    return json as TransportRequirementsType;
}

export function TransportRequirementsTypeToJSON(value?: TransportRequirementsType | null): any {
    return value as any;
}

export function TransportRequirementsTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): TransportRequirementsType {
    return value as TransportRequirementsType;
}